Load user-data as YipConfig files
This commit tries to load user-data from providers as YipConfig files,
if the parsing fails it does nor raise any error. User-data is stored to
the provided path. The file is named `userdata.yaml` if it could be
parsed as a YipConfig structure otherwise it is just called `userdata`.

In addition this commit also includes logic to apply the authorized keys
if the provider extracts a `ssh/authorized_keys` file.

Signed-off-by: David Cassany <dcassany@suse.com>