Use shlex.Split to split dotNotation files
Currently if we try to load a dotNotation file and that file has spaces
in them, i.e. commands in quotes, it will panic and fail to load.

This patch changes the split of the file to use shlex which is context
aware of quotes and preserves that as one item insteaed of splitting it
further thus avoiding strange parsing like ">:true" which gojq fails to
parse completely.

Alos adds some tests for malformed files and half malformed files

Signed-off-by: Itxaka <igarcia@suse.com>