layout: resize the filesystem after expanding it
This patch runs the proper commands to resize the underlying fs after
growing a partition, otherwise the fs will still have the older size
after the growing.

This path also modifies the CmdMock to allow for using regexp for
comparing the commands as there may be variable things that are out of
our control and we cnanot match them exactly.

Also adds tests for the new functionality.

Signed-off-by: Itxaka <igarcia@suse.com>