layout: Check for valid label on xfs systems
xfs filesystems cannot have labels longer than 12 chars.

This patch adds a check at the start of the layout plugin before
touching the disk in order to check that the label is no longer than 12
characters if we are modifying an xfs partition.

Adds tests for it and the correct working of the check on non-xfs fs.

Also modifies the tests a bit so the mocking commands can be more
dynamic and sets vars for the deviceLabel and partition label so they
are all the same across all the tests.

Signed-off-by: Itxaka <igarcia@suse.com>