Improve the partition reload detection
Instead of directly doing the retry on the NewPart call, make it on each
fo the functions that are sensible to delay. Not only its better as they
are independent but its easier to see where it did fail AND if something
else is also calling those functions, they get the retry as well

Also revert to using partprobe as blockdev fails during early boot (i.e.
rootfs.after)

Signed-off-by: Itxaka <igarcia@suse.com>