Add default values after loading them from the env
If for any reason the /etc/default/useradd doesnt have any of the
options that we use (HOME, SHELL) after setting the defaults it gets
overwritten via the godotenv.READ so no matter what you set as the
default in usrDefaults, it gets overwritten.

This makes sure to first load the default values from the files and then
check if they are empty adn fill them with a default value so those
values can be used afterwards.

This was causing the user folder to be created under / on debian systems
as they do not have that default HOME value in the /etc/default/useradd.
Well, its there but its commented out.

Signed-off-by: Itxaka <itxaka@kairos.io>