Remove force name for cloud init configs
Otherwise, in files which have no naming for example:

```
stages:
  default:
    - commands:
      - echo "01"
```

The name of the whole thing will be autogenerated with a fixed string
and it will collide if there is more than one op across different files.

Thus it would somehow miss the files in between the first and the last.
So if you had 5 files it would ignore the 2,3,4 in teh middle and only
run the first and last.

This is due the construction fo the op name which would be like:
`config.Name.Op_number`

Probably because this files are coming from the cloud init loader, the
config name will never change, thus generating ops with the final name
for all of them:

`Cloud init.0`

This patch removes the forced set of the config name to `Cloud init` so
it can be autogenerated based on the file name the op is coming from
thus avoiding any collision in names.

Signed-off-by: Itxaka <itxaka@kairos.io>